refactor: extract trigger e2e tests (#753)
